codigo/requirements.txt
# Dependências para o Scanner de Códigos de Barras
# Instalar com: pip install -r requirements.txt

# Leitura de códigos (wrapper do ZBar)
pyzbar>=0.1.9

# Processamento de imagem e câmera
opencv-python>=4.8.0

# Manipulação de imagens
Pillow>=10.0.0

# Arrays numéricos
numpy>=1.24.0

# Geração de QR codes (para testes)
qrcode>=7.4.0
